requirements.txt
annotated-types==0.7.0
cachetools==5.4.0
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
distlib==0.3.8
filelock==3.15.4
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
identify==2.6.0
idna==3.7
iniconfig==2.0.0
nodeenv==1.9.1
packaging==24.1
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.6.0
pyparsing==3.1.2
pytest==8.3.1
pytest-mock==3.14.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
ruff==0.5.2
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.3
